pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "decision-space-harness"
version = "0.1.0"
description = "Phase 1A implementation of the decision space harness"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "pydantic>=2.12.5",
  "orjson>=3.11.7",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.2"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
